requirements.txt
mlx==0.31.1
mlx-embeddings==0.1.0
datasets>=3.4.0
fastapi==0.135.3
uvicorn==0.42.0
jinja2==3.1.6
pillow>=10.4.0
pillow-heif>=0.21.0
tqdm>=4.67.0
huggingface_hub>=0.34.0
